add remove fd server code,location manger StartServices,fix RSSI-distance model

+++ app/src/main/java/cuhk/cse/cmsc5736project/models/Beacon.java
@@ -63,7 +63,7 @@ public class Beacon implements Serializable
         return rssi;
     }
 
-    public double calDistance(int power)
+    public double calDistance(double power)
     {
         return Math.pow((one_meter_power/power),(1/one_meter_power));
     }
